Refactor BaseE2ETestCase to extend BaseTestCase and remove BaseTestCaseWithSqlDatabaseAccess
